% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\docType{data}
\name{bnfs}
\alias{bnfs}
\title{Define the BNF levels}
\format{
An object of class \code{list} of length 4.
}
\usage{
bnfs
}
\description{
Define the labels of the BNF (in order of hierarchy) with the columns
that are used to aggregate
}
\keyword{datasets}
