% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\docType{data}
\name{breakdowns}
\alias{breakdowns}
\title{Define the breakdowns}
\format{
An object of class \code{list} of length 8.
}
\usage{
breakdowns
}
\description{
Define the labels of the breakdowns (in order of hierarchy) with the columns
that are used to aggregate
}
\keyword{datasets}
