% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fa_to_png_to_datauri}
\alias{fa_to_png_to_datauri}
\title{fontawesome save to datauri
taken from https://jkunst.com/highcharter/articles/fontawesome.html}
\usage{
fa_to_png_to_datauri(name, ...)
}
\arguments{
\item{name}{fontawsome name}

\item{vars}{Grouping variables}
}
\value{

}
\description{
fontawesome save to datauri
taken from https://jkunst.com/highcharter/articles/fontawesome.html
}
