% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{format_data_raw}
\alias{format_data_raw}
\title{Format data-raw table}
\usage{
format_data_raw(df, vars)
}
\arguments{
\item{df}{Dataframe}

\item{vars}{Grouping variables}
}
\value{

}
\description{
Deal with factors and sort table.
}
