% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\docType{data}
\name{geographys}
\alias{geographys}
\title{Define the geographys}
\format{
An object of class \code{character} of length 4.
}
\usage{
geographys
}
\description{
Define the labels of the geographys (in order of hierarchy) with the columns
that are used to aggregate (if there are two columns then the second is the
code)
}
\keyword{datasets}
