% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{statistical_disclosure_control}
\alias{statistical_disclosure_control}
\title{SDC}
\usage{
statistical_disclosure_control(df, col, type, original_col = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{col}{Column to apply it to}

\item{type}{c("count", "percentage")}

\item{original_col}{Original column (if the new column is a percentage and we
need to calculate it)}
}
\value{

}
\description{
Statistical Disclosure Control
}
