% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\docType{data}
\name{geographys}
\alias{geographys}
\title{Define the geographys}
\format{
An object of class \code{list} of length 4.
}
\usage{
geographys
}
\description{
Extract them from the breakdowns.
}
\keyword{datasets}
